all integration specs need to be in the same file
Since you can't spawn and terminate many instances in the same java test process.
